Refactor notebook code and API functions, add new components and update dependencies
